'_mod' removed
'_mod' termination in module filenames isn't a requirement anymore
