+++ AutomatiK files/bot.py
@@ -76,7 +76,7 @@ class Loader:
             module_extension = os.path.splitext(i)[1]
 
             if module_extension == ".py" and module_name != "__init__":
-                class_name = module_name.split("_")[0].capitalize()
+                class_name = module_name.capitalize()
                 try:
                     # noinspection PyPep8Naming
                     Klass = getattr(importlib.import_module(f"modules.{module_name}"), str(class_name))
